Books Controller authorization updated, menu item 'Locations' was removed.
Users delete view is updated so that only the username appears in the delete screen.

Group delete operation updated so that a group with at least one related user cannot be deleted.

first name + last name null exception in user update service fixed.

first name + last name required warning is added to create service.

minimum password length set to 8 in user register view.

user field score is set to nullable, and its default .net error message is fixed.

item.Group values are changed to item.GroupTitle to solve the group column being empty problem in users index view.

User count field is added to groups index view, and related users list was added to groups details view.

required error messages instability in user create and update services were fixed.

+++ MVC/Controllers/GroupsController.cs
@@ -17,18 +17,21 @@ namespace MVC.Controllers
     {
         // Service injections:
         private readonly IService<GroupRequest, GroupResponse> _groupService;
+        private readonly IService<UserRequest, UserResponse> _userService;
 
         /* Can be uncommented and used for many to many relationships, "entity" may be replaced with the related entity name in the controller and views. */
         //private readonly IService<EntityRequest, EntityResponse> _EntityService;
 
         public GroupsController(
-            IService<GroupRequest, GroupResponse> groupService
+            IService<GroupRequest, GroupResponse> groupService,
+            IService<UserRequest, UserResponse> userService
 
         /* Can be uncommented and used for many to many relationships, "entity" may be replaced with the related entity name in the controller and views. */
         //, IService<EntityRequest, EntityResponse> EntityService
         )
         {
             _groupService = groupService;
+            _userService = userService;
 
             /* Can be uncommented and used for many to many relationships, "entity" may be replaced with the related entity name in the controller and views. */
             //_EntityService = EntityService;
@@ -149,14 +152,21 @@ namespace MVC.Controllers
 
         // POST: Groups/Delete
         [HttpPost, ValidateAntiForgeryToken, ActionName("Delete")]
-        [Authorize(Roles = "Admin")] // Only authenticated users with role Admin can execute this action.
-                                     // Overrides the Authorize defined for the controller.
+        [Authorize(Roles = "Admin")]
         public IActionResult DeleteConfirmed(int id)
         {
-            // Delete item service logic:
+            var hasUsers = _userService.List().Any(u => u.GroupId == id);
+
+            if (hasUsers)
+            {
+                SetTempData("This group cannot be deleted because it has related users!");
+                return RedirectToAction(nameof(Index));
+            }
+
             var response = _groupService.Delete(id);
-            SetTempData(response.Message); // set TempData dictionary to carry the message to the redirected action's view
-            return RedirectToAction(nameof(Index)); // redirect to the Index action
+            SetTempData(response.Message);
+
+            return RedirectToAction(nameof(Index));
         }
     }
 }
